Upgrading to new flutter
Use Float32Array to utilize data directly
Utilize new Camera preivew pause functionality
Use compact views to try to save iOS.
Minor import cleanup

+++ lib/tflite/pytorch.dart
@@ -4,7 +4,7 @@ import 'package:flutter/services.dart';
 
 class OutputTensor {
   final Int64List shape;
-  final Float64List data;
+  final Float32List data;
 
   OutputTensor(this.shape, this.data);
 }
